[guest,tests/rust_guests] change guest library to conform to new APIs in common library and use CGM
The host now defines a single undifferentiated memory region—the custom guest memory region (or, CGM)—that
the guest can address as it sees fit. This commit changes the guest library to leverage that and perform
said address using the APIs defined in the common library.

+++ src/tests/rust_guests/simpleguest/src/main.rs
@@ -40,7 +40,7 @@ use hyperlight_common::flatbuffer_wrappers::function_types::{
 use hyperlight_common::flatbuffer_wrappers::guest_error::ErrorCode;
 use hyperlight_common::flatbuffer_wrappers::guest_log_level::LogLevel;
 use hyperlight_common::flatbuffer_wrappers::util::get_flatbuffer_result;
-use hyperlight_common::mem::PAGE_SIZE;
+use hyperlight_common::PAGE_SIZE;
 use hyperlight_guest::entrypoint::{abort_with_code, abort_with_code_and_message};
 use hyperlight_guest::error::{HyperlightGuestError, Result};
 use hyperlight_guest::guest_function_definition::GuestFunctionDefinition;
@@ -557,7 +557,7 @@ fn test_guest_panic(function_call: &FunctionCall) -> Result<Vec<u8>> {
 fn test_write_raw_ptr(function_call: &FunctionCall) -> Result<Vec<u8>> {
     if let ParameterValue::Long(offset) = function_call.parameters.clone().unwrap()[0].clone() {
         let min_stack_addr = unsafe { MIN_STACK_ADDRESS };
-        let page_guard_start = min_stack_addr - PAGE_SIZE;
+        let page_guard_start = min_stack_addr - PAGE_SIZE as u64;
         let addr = {
             let abs = u64::try_from(offset.abs())
                 .map_err(|_| error!("Invalid offset"))
